Initial commit, no code yet.
While I understand this is meant to be for an elixir position, I
decided to use erlang for speed, as I have an idea what libraries and
syntax I would use. The core principles are the same, however, and the
code should be readable even by those with only passing familiarity.