Extended tests to add datastore touch count.
Again, normally I'd skip this commit, but it shows how the project is
evolving.

The poller resaonably only fetch one user at a time. The github api
will reject requests if we poll too much. So, we need a way to know which
user(s) to poll for next. We can keep track of that in the same datastore,
just need to extend it. Thus, new test, new api, new running.